add disk read function in c and fix bug

+++ src/include/io.h
@@ -3,7 +3,7 @@
 
 #include "type.h"
 
-u16 inb(u8 port);
+u8 inb(u16 port);
 u16 inw(u16 port);
 
 void outb(u16 port, u8 val);
